Upgrade log4j
The version we had was so old that it is unlikely to have had the
recent, well-publicized vulnerabilities, but it will make security
folks calmer to have the latest, patched version.

Plus it was somewhat lower-hanging fruit out of the many libraries
that need upgrading since it has very few dependencies in the
application.