Remove JGoodies look and feel options
The library is no longer free so we cannot upgrade it to current.
It's only used for non-macOS, non-Windows users.  Other users are
likely to be able to obtain and install their own look and feel
options, and even more likely to have a Java or OpenJDK provider
who has already done so for them.  Metal is better than it used to
be and Nimbus is now included with the JRE and can be relied on to
be always present.

So make Nimbus the default for non-macOS, non-Windows but still
allow users to choose other options available on their systems.