Get rid of SmoothWindows
We use exactly one method from one class, and the rest was cruft.
Deprecated cruft as it all references com.sun.java.swing.plaf, which
is no longer a public API and is likely to be removed entirely.
This allows us to remove a few dozen unused files and 1500+ lines
of dead code.  And we can stop suppressing the warning about using
methods from the internal API.