Attempt to get error dialogs in front
This is a somewhat desperate hack that mostly keeps error dialogs
from appearing behind the windows that spawned them by putting
them below that window if there is a parent, or centered near the
top of the screen if there isn't.  All of this brought to you by
the fact that toFront(), setComponentZOrder(), requestFocus(),
and requestFocusInWindow() all have no visible effect on JDialog
components.

A better solution would likely involve a complete redesign of the
dialog system using JLayeredPane.