We have this thing called version control
It's much easier for people interested in building WordHoard for
themselves to just clone the repository rather than unzipping a
bunch of different zip files and manually putting them in place.
It also makes deployment a bit simpler.

The pre-built database files, the data file containing the texts,
and the library dependencies are not in the repository, so we still
provide those.