Fix setting Nimbus on non-macOS, non-Windows
Setting the default look and feel on these systems was only checking
against the list of additional options installed with WordHoard
but we don't actually have any of those any more.  So make it look
through all of the available options, including those that came with
the system, when verifying the look and feel we are setting.

Without this, Nimbus was not getting successfully set.