Indicate macOS program name in jnlp file
We set this programmatically in our own start-up code and that
works when running locally.  icedtea sets it to "Boot" willy-nilly
before our code gets a chance to set it, and while it should honor
the properties in the jnlp file, it doesn't.  But put it in there
anyway in case that ever gets fixed, and put it before other
properties in case the order makes a difference.