Implement regular expressions in queries
Previously, WordHoard extensively customized Hibernate itself to
provide this.  As of Hibernate 5, regular expressions are *still*
not available out of the box, but there is a less invasive
mechanism to provide a custom function based on the RLIKE operator
in MySQL/mariadb.  The function is made available by providing a
custom dialect that subclasses the MySQL dialect and then
referencing the custom dialect everywhere a dialect is specified.

The return value of the function cannot be treated as a boolean
in HQL but must be checked against a value, which means doing
some surgery to the query builder so that regular expressions can
be handled without breaking other query types.