Must specify database for user data tables
With Hibernate 5, the server will attempt to find these tables in
the wordhoardserver database, but that's not where they are, so we
have to explicity specify the wordhoad database.

Also, when producing wordcounts during wordset creation, do not
garbage collect before the counts are used as that can make them
disappear.