Set tree selection for Java 11
We don't use TreeSelectionListener because it might interfere with
the drag and drop we have enabled on the table of contents and
search results trees.  With Java 8, selecting a tree node selected
that node without further intervention, but with Java 11, the node
selection does not change on mouse down, so we have to set the
selection to the click coordinates explicitly.