Allow Java 11 in jnlp file
According to the JNLP API spec:

    A version string is a list of version-ids separated with spaces.
    Each version-id can be postfixed with a '+' to indicate a
    greater-than-or-equal match, a "*" to indicated a prefix match,
    or have no postfix to indicate an exact match.

So we specify 11.0 first and 1.8 as a fallback.  We're implicltly
disallowing Java 9 and Java 10 as these are not LTS releases.
Hopefully we can add 17.0 to the list at some point, but that will
be another investigation.

While we're tweaking the j2se, bump the memory settings a bit as
nothing ever takes less memory than it did thirteen years ago and
it is possible with some very complex analyses to run out of heap.