Make the move from javax to jakarta
This will be required before going to Hibernate 6.x.  It ended up
adding several new dependencies and a lot trial-and-error to find
versions of those libraries that will work together and also still
allow compilation to a Java 8 target, as most of them have moved
on to something else in their current versions.