Include drop as well as create in schema export
This was harder than it should be because if you tell it to override
an existing file, it overrides its own drop script to do the creates
even if you do them both with a single call to execute().  So just
use basic Java facilities to delete the file if it exists before
script generation.