Convert from XDoclet to JPA annotations
The doclet approach has been on its way out for a long time, with JPA
annotations being the recommended alternative despite many deficiencies (see
below). This commit will leave the old annotations and the ability to generate
the .hbm.xml files using XDoclet in place, to be removed in a subsequent
commit. This removal will be necessary in order to compile with Java 11 or
later, which is in turn necessary to be able to upgrade to Hibernate 6.x or
later.

The schema produced here is 99% the same as the schema generated the old way,
with column order being the main dealbreaker that prevents it from being a
drop-in replacement. For now we'll keep the old misc/schema.ddl in place along
with the new misc/wordhoard_schema.ddl. Supposedly Hibernate 6.x will provide
a way to control column order so we may be able to get the correct order back
after that upgrade.

Notable problems encountered converting to Hibernate/JPA annotations:

- JPA annotations are much more verbose, requiring lots of repeated code.
- No provision for declaring an index on a column at the column level (must be
  on the table declaration).
- No robust conversion tools for migration from either doclets or .hbm.xml
  (HBM2Java is sometimes helpful but always incomplete and often wrong).
- Cannot infer primary key for collection or join table from a List object,
  thus necessitating the use of columnDefinition, which is non-portable and
  deprecated.
- No provision for specifying a prefix for an embedded class, but instead
  requires repeating each embedded field in an @AttributeOverride
  declaration.
- No ready-made build tools to produce either .hbm.xml files or DDL from the
  annotations.
- Every getter method that is not to be persisted must be marked explicitly
  with @Transient, even superclass methods whose overrides you've already
  marked @Transient.
- Sometimes chokes on a getter method with no corresponding setter,
  necessitating the creation of an empty setter.
- Generates create table statements with columns in alphabetical order and no
  way to specify non-default order (supposedly will change in Hibernate 6.x);
  obviously defaulting to anything other than declaration order is a bug.