Bootstrap from annotations
Primarily a matter of loading annotated classes, but a couple of
other things go with it:

1.) Make MetricalShape embeddable rather than persistent

This means we will no longer get a metricalshape table, but we never
had any data in it so it was probably always a mistake to persist it.

This fixes a hard-to-diagnose error in which Hibernate claimed there was
a duplicate ID field in the Word class.  It was really complaining about
the fact MetricalShape had its own identifier yet was also a property of
Word.  Since we were always using MetricalShape as if it were embeddable,
making it genuinely so and removing its id property seemed the best path
to take.

This also allows us to back out a workaround where were flip-flopping
between treating MetricalShape as a class and as a string.

2.) Remove Hibernate taglets

We're now generating DDL from annotations, and also bootstrapping from them,
so we no longer need them. That should be one obstacle down in getting ready
for Java 11+.