Correctly evict by class
The "Clear Caches" button under System Information was throwing this:

java.lang.IllegalArgumentException: Non-entity object instance passed to evict

It turns out the evict method on the session takes an entity object, not a class,
so iterating over our list of persistent classes and evicting each class wasn't
giving it what it expected.  There are reports that the exception was not thrown
before the Hibernate 4.x era, so this probably has not worked since inception
but failed silently when we were using Hibernate 3.x.

Luckily we can obtain the cache itself from the session factory and the cache
does have an evict method that takes a class and clears "the data for entities
of the specified class (and its subclasses)."