Replace source and target with release; build with Java 11+
Now that the doclets are gone, we can compile with Java 11. But it
turns out we can also compile with Java 17 so go ahead and start
using that.

In Java 11 and later, target, source, and bootstrappath are all
deprecated and we can use the much simpler release parameter on the
compile. Aim at Java 8 so we can run with 8, 11 or 17.